KUDU-2416: Fix PartialRow.setMin and add a unit test
* Fixes INTEGER.MIN_VALUE being used for LONG colums
  that could lead to incorrect partition prunning.
* Adds a unit test for all types.

Reviewed-on: http://gerrit.cloudera.org:8080/10191
Tested-by: Grant Henke <granthenke@apache.org>

+++ java/kudu-client/src/main/java/org/apache/kudu/client/PartialRow.java
@@ -1024,7 +1024,7 @@ public class PartialRow {
         break;
       case INT64:
       case UNIXTIME_MICROS:
-        addLong(index, Integer.MIN_VALUE);
+        addLong(index, Long.MIN_VALUE);
         break;
       case FLOAT:
         addFloat(index, -Float.MAX_VALUE);
